+++ packages/deploy-config-sub-generator/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @sap-ux/deploy-config-sub-generator
 
+## 1.0.31
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/abap-deploy-config-sub-generator 1.0.25 → 1.0.26
+- @sap-ux/fiori-generator-shared 1.0.14 → 1.0.15
+- @sap-ux/inquirer-common 1.0.15 → 1.0.16
+- @sap-ux/odata-service-inquirer 3.0.17 → 3.0.18
+- @sap-ux/cf-deploy-config-sub-generator 1.0.19 → 1.0.20
+- @sap-ux/deploy-config-generator-shared 1.0.14 → 1.0.15
+
 ## 1.0.30
 
 ### Patch Changes