+++ packages/inquirer-common/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @sap-ux/inquirer-common
 
+## 1.0.16
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/telemetry 1.0.12 → 1.0.13
+- @sap-ux/fiori-generator-shared 1.0.14 → 1.0.15
+
 ## 1.0.15
 
 ### Patch Changes