+++ packages/odata-service-inquirer/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @sap-ux/odata-service-inquirer
 
+## 3.0.18
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/telemetry 1.0.12 → 1.0.13
+- @sap-ux/fiori-generator-shared 1.0.14 → 1.0.15
+- @sap-ux/inquirer-common 1.0.15 → 1.0.16
+
 ## 3.0.17
 
 ### Patch Changes